Elements Colours
Implemented temporary visual feedback for when we change element type by changing spriterenderer.color whenever we swap. Also refactored a small amount of code to only call GetComponent in void Start so it only has to be called once, then used the stored variables in script where it was used before.