Collision Implementation
Created a platforms script to manage lists of objects for each element. It grabs the bools from the player script to check which collision needs to be turned on or off. Probably not the best implementation (potentially quite expensive depending on how many objects we have) but good enough to prototype with for now.

Unfortunately the jump functionality will need to be changed since Unity doesn't allow for more than one tag per gameobject. Possible workaround for this is to create child/parent gameobjects with the separate tags, create a custom script for tags or look for a different approaches to reference specific gameobjects without using tags (e.g. types).

Also created a platform_manager prefab since we'll need a gameobject with that script on in every scene for now.