Projectile Trigger
Added the void OnTriggerEnter2D function to our projectile manager script and added some temporary code for hitting enemies. Same workaround as the collision if we want to add element types to the enemies, made a parent gameobject with the enemy tag to work with. At the moment enemies and the projectile are destroyed when it is triggered.