Improved Double Jump
Set the y velocity to 0 just before we add the jump force meaning that we will always get a reliable distance. Just adding the force meant that it would take into account falling speed so we would jump high if we jumped at the apex, but low if we jump while falling.