Jumping Using Raycast
When we enter a collider now firing a raycast directly down to check we are on top of something to be able to jump. This fixes bugs like being able to continuously jump into the bottom of a platform since before we were just checking for any kind of collision. Need to finalise whether we want to let the player jump/double jump while falling. Added some extra useful debug prints for this too.