Dashing
Added functionality for dashing left and right. May need to revisit to tweak to our needs, at the moment checking for 80% velocity after dashing then stopping so it feels like a short burst of speed. Potentially might want to disable left/right movement while we are dashing. At the moment left and right dash is bound to Q and E but we can look at adding double tapping in future if that's what we want.