Death/Respawn at Checkpoint
Added functions to playerscript to handle dying and respawning. Added a checkpoint class with a trigger to handle where the current checkpoint is and updating in playerscript when the player reaches a new checkpoint. Also created a prefab for our checkpoint so can add sprites or whatever other components we need to that later.