Crouching
Setup crouching to work with a collider, no sprite changes/animations yet but it seems to work as expected. Also changed the layer to ignore raycasts on all our projectile prefabs because there was a bug with the way collision has been setup where if a projectile was fired downwards as the player collided with something, the raycast would hit the projectile and mess up the jump logic.