Destructible Objects
Added functionality for destructible objects when we're earth by using oncollisionenter and oncollisionexit in the player script to check for "melee" range. Added functionality for attack speed with a value that can be edited in the editor. Fixed null exceptions that would happen when the raycast for jump checks hits something without a parent.