Earth Dash/Fast fall
Added earth dash/ "earthfall" function to quickly dash down as earth. Currently bound to the e key when we're earth, adds a changable force. I found if the force was >~40 (at the height of our current level) started having some collision issues because we were going so fast. Enabled continuous collision detection while we're falling just so this doesnt happen, but might be better to just avoid a high force anyway. Also put crouch code into functions to make it a bit more readable, could be better potentially changing functions to take in variables but that can wait until this script is refactored and evaluate if that's even necessary.