Grabbing functionality
Implemented grabbing functionality while we're water when holding E. Also implemented a wall jump that jumps in the opposite direction to the wall we are grabbing. Firing raycasts left and right when we collide to check if there is an object we can grab onto, if there is then detect input every frame for E so we can hold it before we touch the surface. Might be worth changing this to a bool to improve performance that is turned on by the keypress down.