Prototype Script
Set up a "TestPlayerScript" that forces the input of each element's ability when we change element for prototyping. Made it dash in the direction we're moving with some small tweaks and ensured we can't infinitely jump with air. Attached to the player for now, not added it to the prefab.