Fixed Deceleration
Fixed deceleration - wasn't working because I wasn't checking that the value was between 0 and our deceleration when zeroing out the velocity when it was too small so it was just instantly zeroing out the velocity whenever we called the decelerate function. Movement feels a lot tighter now, still needs refactoring and potentially changed to handle forces instead.