Movement Script
Began refactoring movement into a single script rather than handling it in the player script. Altered functions to take in a rigidbody and two floats for acceleration and maximum speed as parameters and handled movement in fixedupdate to avoid different values at different framerates. Made a temporary test player gameobject in the scene to use for testing the scripts I will be refactoring out of the main player script.