CollisionManager and JumpScript
Refactored our jump into a script that handles it separately using fixedupdate. To do this, I began work on the collisionmanager script that handles our functions to detect when we're grounded and colliding with walls. Still need to implement wall checks but works as intended for the grounded checks for now.