Dashing
Refactored dash script, needs to communicate with the movescript for some functionality e.g. which direction to dash in. Could improve functions to be more consistent and some tweaking might be necessary to clean up the gravity management in update, since we are modifying physics in update instead of fixedupdate.