Setup Coyote Time and walljump fix
Added coyote time in the collision manager script. Implementation is a bit clunky but seems to do the job ok, needs more testing to see if it causes any serious issues. The value for the amount of time the coyote time lasts for can be tweaked on the collision manager script in editor. Also fixed a bug where walljump would add the force to any positive upward velocity if the player can spam walljump by zeroing out the velocity before we walljump.