Added functionality for pre-jumping
Sending raycasts down while we're falling, similar to grounded checks to see if we're going to hit the ground soon. Depending on the length we want the raycast, we can be more or less generous. Can be changed with the preJumpDistanceModifier value in editor. Also added debug rays when using gizmos to see the raycasts to make it easier to play about with and get a workable value. Just enable the "Draw Pre Jump Raycast Debug" in editor and enable gizmos when the game has launched in the top right corner to see them.