Pre-jump fixes
Fixed a couple of bugs with pre-jumping - one where jumping next to a platform slightly higher would trigger pre-jumping as soon as we landed despite not pressing space. Seems to have been because pre-jump wasn't getting made false correctly so setting that when we're grounded every frame now. Another bug when we were earth we could pre-jump if we timed the jump correctly so checking if we're earth too now before running the raycasts and clearing the preJump bool if we are earth.