Saving Scenes in player data
Changed the save system to handle saving the scene the player is in when saving the player data. We can turn this off when testing individual scenes by unchecking the bool in PlayerState in editor. Created a quick checkpoint in our samplescene to check this is working. Turned it off on the scenes we have atm and defaulted it to off in script.