Controller Functionality
Made a start with controller functionality in script just to get a base input. Potentially we want to use the input package manager to handle this, but for now set it up in the legacy input manager and referenced the specific joystick binds in script. The left and right movement feels VERY clunky at the moment because any movement across the X axis triggers the maximum acceleration in whichever direction - will need to be revisited to get this to travel a max speed based on the position of the left stick.