Scene Portal
Added a scene portal script and made a prefab that uses a trigger collider to detect if the player is inside it, and if it is to load the scene specified in the editor on the script as a string. Note that build settings will have to include each scene we want to use for this in File>Build Settings.