Modifications to earth dash
Made a small change to the earth dash script so that we only zero out the y velocity when we're travelling upwards. If it's 0 or below, we just add the force to the current velocity. May still want to look at changing/turning off linear drag (maybe at runtime) since we're turning gravity off and it can have a slowing effect but it might only be noticeable at long distances so might not even be necessary.