Earthdash while on platforms fix
Found a bug where the platform's collision would not get disabled fast enough when earth dashing from on top of a platform. Fixed this by storing the game object we are grounded on in a variable and checking if it is not an earth platform when earth dashing. If it isn't an earth platform, its collision is disabled just before the force for the earth dash is applied.