Earthdash incorrectly resetting gravity
Found a bug where our fire dash would be falling if we dashed from a surface rather than in the air. Found the cause was that earth dash was resetting gravity every frame we were grounded because of something wrong with the if statement to check if a dash had ended.