Breakable Platforms and Platform parenting
Added functionality for platforms to be breakable, which disables the collision and after a time we specify in script it is destroyed (in case we want to add some sort of destroyed animation and have it fade out). Also made our platforms function correctly when our player is on it by simply having the platforms parent the player's transform when they enter the collision and unparenting when the player exits the collision.