Breaking Platforms from dashing from a specific height
Added functionality in platform script and earthdash script to calculate the distance we travelled in our earth dash. We can change in editor per platform what value we want to the platform to be breakable from - from a small amount of testing it felt like ~6.0 was a reasonable height to require a double jump to consistently break it.