Health/Damage + Dying/Respawning
Set this up pretty simply where if you die you go to the last checkpoint you were at. Made a prefab for a damage volume which takes the specified damage in the script (can be changed in editor) when collided with. After we take damage we check if we're less than 0 health and if we are we respawn at the last checkpoint. May need to change timings for animations or anything else but the basic functionality is there with logging to console.