Health script changes
Changed this script to be less dependent on other scripts for its functionality so this can be reused on whichever objects we like in future to have health/damage and respawn functionality. It just needs to be fed a checkpoint value by another script (in our case the checkpoint script) to respawn at.