Camera Shake
Messed around today a bit with the camera script and decided to add a camera shake when we landed from above a certain height with our earth dash. Made the shake duration editable within the editor but 0.1 seems to be a pretty good duration for now. Did some brief testing in the samplescene with it.