Reworked Pre-jump
Reworked it so it is based off a timer instead of unnecessary raycasts. Also found a bug if we earth dash and air jump straight after sometimes gravity is not reset correctly. Changed some things to fix it but it's quite rare, will have to see if it crops up again.