Camera Shake refining
Noticed that the way I'd set camera shake up it would stop shaking as soon as we changed element which felt weird. Fixed this by only disabling the camera shake once the duration was finished. Also noticed that the coyote time wasn't working properly still because I was incrementing the frame buffer before checking it was incremented meaning that it just ran through every time. Moved the increment after the check to fix it.