Platform parenting fixes
Changed the way we parent the player's transform to a platform since we're using the same script for obstacles. Now we only parent if the player is grounded on the platform. This might have some implications for if we want to use oscillating platforms with water form but can make those changes when necessary if we decide to make levels with something like that.