Animation Controller Implementation
Began implementing more animations into a water state machine. Feels a bit messy at the moment and there's sure to be lots of bugs so loads more work needed on this but the wall sliding and jumping animations seem to work really well. Setting flags in relevant scripts and passing them to the animation controller.