Animation Bug Fixes
Set up all but 2 of the animations in the controller and setting the values in script in the appropriate places (may need refining). Did some testing and everything seems to be working as expected.

Also changed the direction the player is facing after wall jumping - it was pointing the player towards the wall they jumped from, so dash would face towards that too. Changed it so that the player faces the direction they are jumping in, seems to be more intuitive with the animations.