Reimported again to change the jump again.
I split the jump into ascent, desecnt and landing. We might not need them all but this way we can script the game to play the landing anim so we have a nice squash. You will only really see this naim if you are static I think. oterhwise you will have to interupt the anim to change to a run.