Elements UI
Started work on this, mainly editor work with a small amount of scripting to access the variables we need to see which element we've changed to.. If the UI is added to a new scene, the player game object must be added to the elements bar for the playerstate and dashscript variables in editor. Simplistic UI at the moment just to tell the player which element they are, need to implement cooldown UI for dash and show if air jump is available + health UI and implementation for that system.