Platform Changes
Lots of changes to platforms; made tile and platform collision layers in project settings to avoid unwanted collisions. Moved the oscillating functions over to update so that the timers would work correctly, but now using a speed * time.deltatime to make it consistent regardless of fps - NOTE: This means all oscillating platforms as of now are very slow and will need their values readjusting. Also noticed that the oscillating functions weren't placing the platforms in their final position if their next frame position was > than the final position. I believe this is what was causing the de-sync of platforms with the same properties. Hopefully all these changes combined will make iteasier in the long run to design levels with oscillating platforms.