Lives/health System and Game over screen
Implemented lives functionality and a game over screen for when the player hits 0 lives that allows the level to be restarted or the player to go back to the main menu. This required changing some of the save functionality, so now we only save as soon as we reach a new level.