Changing Alpha of platforms when they're collision is disabled
Reduced the alpha on platforms that have their collision disabled. Also tweaked the platform script to use the OnBecameVisible and OnBecameInvisible methods to only enable/disable collision when platforms are visible by our camera. This also required some implementation within the earth dash function since we're disabling collision there instead of the regular function when we earth dash from directly on top of a platform of a different element.