Tooltip Trigger & Tooltip
Added a simple text object in our UI canvas prefab (feel free to tweak text/position) and made a quick script for the ability to trigger it and saved a prefab. Basically all the prefab is is a box collider to trigger the tooltip - the script attached to it allows us to change the text of the tooltip, change if we'd like to trigger it more than once and for how long we'd like it to appear. For now just make sure we drag the tooltip game object from the UI canvas into the "Tooltip" variable in editor, and the player game object to the "Player". Would still like to come back to it to make it handle some error cases.