Health UI reworked
Reworked the health UI to use an animation controller to hook up the new UI. The sprites for the anims will need to be a little bigger but this works as expected at the moment. Had to change the UI canvas to use screen space - camera and use the main camera as the render camera to use sprites on the canvas - this will need to be done for each level. This might cause some issues down the line  so may need to revisit this and change the way this will happen if issues arise.