Scene transition implementation
Implemented scene transitions, a simple fade out then back in for now we can change eventually if we want. Added a canvas called "LevelLoader" as a prefab that has been added to levels 1-5. The scene portal objects need to be assigned the LevelLoader object in editor to work now since they use it as an object to access the transition function.