Some level and animation changes
Checked out our health animation bug and it seems to be because we have some overlapping damage volumes. I've taken one out and of level 2 that was clipping with the other ones but if the player drops so that they hit both volumes the bug can still occur. Might need to change the prefabs we use so it's just one long volume where possible but if for design we want separate damage volumes close to each other I'll need to add some checks in script to make sure we only take damage once even if we hit two in the same frame.